+++ pyproject.toml
@@ -429,7 +429,7 @@ unfixable = [
   "D", # ignore all docstring lints in tests
 ]
 "ci/release/verify_release.py" = ["T201"] # CLI tool that prints stuff
-"ibis/expr/selectors.py" = ["UP007"]
+"ibis/selectors.py" = ["UP007"]
 
 [tool.black]
 line_length = 88
